7 outs works in Logic; remove an incorrect bus assumption
The six sines 7 outs had improperly assumed that busses
were contiguious not just in stereo pair but also across
the busses. WIth CLAP and VST3 this happened to be true
but in AU it is not. So adjust multi-out bus addressing
code appropriately.

This defacto allows the 7 out versoin to work in logic

CLoses #186